requirements.txt
gspread==6.1.2
google-auth==2.29.0
requests==2.31.0
pytz==2024.1
pandas==2.2.2